各Viewの調整
